# perf: improve performance of `project.get_contract()`