### What I did

The performance of getting contracts from projects is bad because it pointlessly always reloaded the manifest. Mistake! Fixed!

For testing, using a MASSIVE PROJECT.

With change:

```
In [1]: %time project.get_contract("MyContract")
CPU times: user 116 ms, sys: 12 ms, total: 128 ms
Wall time: 132 ms
```

Without change:

```
In [1]: %time project.get_contract("MyContract")
CPU times: user 1min 19s, sys: 2.88 s, total: 1min 22s
Wall time: 1min 22s
```


Improvement≈99.84%

So, the performance improvement is approximately 99.84%.

### How I did it

delete code.
test.

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated